Added Articles UI and ArticleController

+++ src/main/java/com/ag04/Feeddit/repositories/ArticleRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface ArticleRepository extends CrudRepository<Article, Long> {
+public interface ArticleRepository extends CrudRepository<Article, Integer> {
 
     public List<Article> findByUserUsername(String username);
 }
